+++ ddr/lib/ddr-blobgen/java/genSuperset.cpp
@@ -222,10 +222,10 @@ SupersetFieldVisitor::visitType(TypedefUDT *type) const
 	}
 
 	/* Get field pointer/array notation. */
-	for (int i = 0; i < type->getPointerCount(); i += 1) {
+	for (size_t i = 0; i < type->getPointerCount(); i += 1) {
 		*_pointerTypeBase = "*" + *_pointerTypeBase;
 	}
-	for (int i = 0; i < type->getArrayDimensions(); i += 1) {
+	for (size_t i = 0; i < type->getArrayDimensions(); i += 1) {
 		*_pointerTypeBase = *_pointerTypeBase + "[]";
 	}
